+++ src/main/java/models/entity/Route.java
@@ -5,8 +5,8 @@ import java.util.Objects;
 
 public class Route {
     private Integer routeID;
-    private Double[] startMarker;
-    private Double[] finalMarker;
+    private Float[] startMarker;
+    private Float[] finalMarker;
     private Integer length;
     private Integer orderID;
 
@@ -18,19 +18,19 @@ public class Route {
         this.routeID = routeID;
     }
 
-    public Double[] getStartMarker() {
+    public Float[] getStartMarker() {
         return startMarker;
     }
 
-    public void setStartMarker(Double[] startMarker) {
+    public void setStartMarker(Float[] startMarker) {
         this.startMarker = startMarker;
     }
 
-    public Double[] getFinalMarker() {
+    public Float[] getFinalMarker() {
         return finalMarker;
     }
 
-    public void setFinalMarker(Double[] finalMarker) {
+    public void setFinalMarker(Float[] finalMarker) {
         this.finalMarker = finalMarker;
     }
